Modularize with buildSchema and graphqlHTTP.
