+++ server/graphql/resolvers/index.js
@@ -1,0 +1,7 @@
+const postResolver = require('./post.resolver');
+
+const resolvers = {
+  ...postResolver
+}
+
+module.exports = resolvers;
