+++ server/graphql/schemas/config.schema.js
@@ -1,0 +1,7 @@
+module.exports = `
+type Config {
+  _id: ID!
+  user_ud: ID!
+  blog_name: String!
+  author_name: String!
+}`;
